refactor: Moved testing utility up to root as it's gonna be shared
